+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.2.5 [unreleased]
 
+- `pgsync group1` is now shorthand for `pgsync groups group1`
 - Added `--schema-only` option
 - Added `--no-rules` option